Chore: refactor integration tests
