+++ tests/core/engine_adapter/integration/docker/_common-hive.yaml
