+++ tests/core/engine_adapter/integration/docker/clickhouse/config.xml
