+++ tests/core/engine_adapter/integration/docker/clickhouse/keeper.xml
