+++ tests/core/engine_adapter/integration/docker/compose.clickhouse.yaml
