+++ tests/core/engine_adapter/integration/docker/compose.mssql.yaml
