+++ tests/core/engine_adapter/integration/docker/compose.mysql.yaml
