+++ tests/core/engine_adapter/integration/docker/compose.postgres.yaml
