+++ tests/core/engine_adapter/integration/docker/compose.spark.yaml
