+++ tests/core/engine_adapter/integration/docker/compose.trino.yaml
