+++ tests/core/engine_adapter/integration/docker/init-metastore-db.sql
