+++ tests/core/engine_adapter/integration/docker/spark/Dockerfile
