+++ tests/core/engine_adapter/integration/docker/spark/conf/hive-site.xml
