+++ tests/core/engine_adapter/integration/docker/trino/catalog/datalake_delta.properties
